Revise to work with new NamedTuple constructor

+++ test/runtests.jl
@@ -22,6 +22,6 @@ it = IteratorInterfaceExtensions.getiterator(pqf)
 ar = collect(it)
 
 @test length(ar) == 25
-@test ar[1] == @NT(n_nationkey = 0, n_name = "ALGERIA", n_regionkey = 0, n_comment = " haggle. carefully final deposits detect slyly agai")
+@test ar[1] == @NT(n_nationkey = Int32(0), n_name = "ALGERIA", n_regionkey = Int32(0), n_comment = " haggle. carefully final deposits detect slyly agai")
 
 end
